+++ src/main/java/itu/framework/backoffice/entities/Trajet.java
@@ -1,22 +1,12 @@
 package itu.framework.backoffice.entities;
 
-import legacy.annotations.Column;
-import legacy.annotations.Entity;
-import legacy.annotations.ForeignKey;
-import legacy.annotations.Id;
-import legacy.query.Comparator;
-import legacy.query.FilterSet;
-import legacy.schema.BaseEntity;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.math.BigDecimal;
 import legacy.annotations.*;
 import legacy.query.Comparator;
 import legacy.query.FilterSet;
 import legacy.schema.BaseEntity;
 import legacy.strategy.GeneratedAfterPersistence;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -42,7 +32,7 @@ public class Trajet extends BaseEntity {
     private LocalDateTime heureArrivee;
 
     @Column(name = "distance_totale")
-    private Double distanceTotale;
+    private BigDecimal distanceTotale;
 
     @Column(name = "ordre_visites")
     private String ordreVisites; // JSON ou CSV: "AERO01,HOT001,HOT002,AERO01"
@@ -131,11 +121,11 @@ public class Trajet extends BaseEntity {
         this.heureArrivee = heureArrivee;
     }
 
-    public Double getDistanceTotale() {
+    public BigDecimal getDistanceTotale() {
         return distanceTotale;
     }
 
-    public void setDistanceTotale(Double distanceTotale) {
+    public void setDistanceTotale(BigDecimal distanceTotale) {
         this.distanceTotale = distanceTotale;
     }
